+++ src/main/java/io/kontur/eventapi/dao/mapper/DataLakeMapper.java
@@ -34,4 +34,6 @@ public interface DataLakeMapper {
     Boolean isNewEvent(@Param("externalId") String externalId,
                         @Param("provider") String provider,
                         @Param("updatedAt") String updatedAt);
+
+    Integer getNotNormalizedObservationsCount();
 }