+++ src/main/java/io/kontur/eventapi/gdacs/client/GdacsClient.java
@@ -1,6 +1,5 @@
 package io.kontur.eventapi.gdacs.client;
 
-import io.micrometer.core.annotation.Timed;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -9,10 +8,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 public interface GdacsClient {
 
     @GetMapping("/xml/gdacs_cap.xml")
-    @Timed(value = "httpClient.gdacs.xml")
     String getXml();
 
     @GetMapping("/gdacsapi/api/polygons/getgeometry")
-    @Timed(value = "httpClient.gdacs.geometry")
     String getGeometryByLink(@RequestParam("eventtype") String eventtype, @RequestParam("eventid") String eventid, @RequestParam("episodeid") String episodeid);
 }