+++ src/main/java/io/kontur/eventapi/typehandler/FeedEpisodeTypeHandler.java
@@ -2,9 +2,9 @@ package io.kontur.eventapi.typehandler;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import io.kontur.eventapi.entity.FeedEpisode;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.ibatis.type.BaseTypeHandler;
 import org.apache.ibatis.type.JdbcType;
-import org.springframework.util.StringUtils;
 
 import java.sql.CallableStatement;
 import java.sql.PreparedStatement;